+++ lib/src/model/anime/anime_item.dart
@@ -29,7 +29,7 @@ abstract class AnimeItem implements Built<AnimeItem, AnimeItemBuilder> {
   String? get synopsis;
 
   @BuiltValueField(wireName: 'type')
-  String get type;
+  String? get type;
 
   @BuiltValueField(wireName: 'airing_start')
   String? get airingStart;
@@ -42,6 +42,15 @@ abstract class AnimeItem implements Built<AnimeItem, AnimeItemBuilder> {
 
   @BuiltValueField(wireName: 'genres')
   BuiltList<GenericInfo> get genres;
+
+  @BuiltValueField(wireName: 'explicit_genres')
+  BuiltList<GenericInfo> get explicitGenres;
+
+  @BuiltValueField(wireName: 'demographics')
+  BuiltList<GenericInfo> get demographics;
+
+  @BuiltValueField(wireName: 'themes')
+  BuiltList<GenericInfo> get themes;
 
   @BuiltValueField(wireName: 'source')
   String get source;